ANA-618: enable dataset.upload(static_csv_file)

+++ gooddataclient/archiver.py
@@ -84,7 +84,8 @@ def write_tmp_zipfile(files):
 
 
 def create_archive(
-    data, sli_manifest, dates, datetimes, keep_csv=False, csv_file=None
+    data, sli_manifest, dates, datetimes, keep_csv=False,
+    csv_file=None, csv_input_path=None
 ):
     """
     Zip the data and sli_manifest files to an archive.
@@ -92,10 +93,14 @@ def create_archive(
 
     @param data: csv data
     @param sli_manifest: json sli_manifest
+    @param csv_input_path: create archive with this
+                           csv data instead of data
 
     return the filename to the temporary zip file
     """
-    if isinstance(data, str):
+    if csv_input_path:
+        data_path = csv_input_path
+    elif isinstance(data, str):
         data_path = write_tmp_file(data)
     elif isinstance(data, Iterable):
         data_path = write_tmp_csv_file(data, sli_manifest, dates, datetimes)
@@ -115,7 +120,8 @@ def create_archive(
         if not csv_file:
             raise TypeError('Keep csv option with no csv file path')
         shutil.move(data_path, csv_file)
-    else:
+    # do not delete the csv file if push of static data
+    elif not csv_input_path:
         os.remove(data_path)
     os.remove(sli_manifest_path)
     return archive
